J'ai pas trop compris ce que t'as fais Lucas

+++ asm.d/srcs/disp_hexlen.c
@@ -3,28 +3,28 @@
 /*                                                        :::      ::::::::   */
 /*   disp_hexlen.c                                      :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: lutsiara <lutsiara@student.42.fr>          +#+  +:+       +#+        */
+/*   By: wahasni <wahasni@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/02/10 16:15:19 by hasni             #+#    #+#             */
-/*   Updated: 2020/02/19 16:41:56 by lutsiara         ###   ########.fr       */
+/*   Updated: 2020/02/22 02:37:56 by wahasni          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "asm.h"
+#include "../includes/asm.h"
 
-void disp_hexlen(int fd, unsigned long size, int len)
+void	disp_hexlen(int fd, size_t size, int len)
 {
-	unsigned char *tmp;
-	int i;
+	unsigned char	*tmp;
+	int				i;
 
 	i = len;
-	if (!(tmp = (unsigned char *)ft_memalloc(i * sizeof(unsigned char))))
-		return;
+	if (!(tmp = (unsigned char*)ft_memalloc(i * sizeof(unsigned char))))
+		return ;
 	while (size && i > 0)
 	{
 		tmp[--i] = size % 256;
 		size /= 256;
 	}
 	write(fd, tmp, len);
-	ft_memdel((void **)&tmp);
+	ft_memdel((void**)&tmp);
 }
